requirements.txt
Flask
requests
icalendar
python-dotenv
pytz # For timezone handling
